src/exchanges/sample/modules/authed/order/listRaw.ts
import { debug } from 'debug'

import { IAlunaExchangeAuthed } from '../../../../../lib/core/IAlunaExchange'
import { AlunaHttpVerbEnum } from '../../../../../lib/enums/AlunaHtttpVerbEnum'
import {
  IAlunaOrderListParams,
  IAlunaOrderListRawReturns,
} from '../../../../../lib/modules/authed/IAlunaOrderModule'
import { SampleHttp } from '../../../SampleHttp'
import { SAMPLE_PRODUCTION_URL } from '../../../sampleSpecs'
import { ISampleOrderSchema } from '../../../schemas/ISampleOrderSchema'



const log = debug('@aluna.js:sample/order/listRaw')



export const listRaw = (exchange: IAlunaExchangeAuthed) => async (
  params: IAlunaOrderListParams = {},
): Promise<IAlunaOrderListRawReturns<ISampleOrderSchema>> => {

  log('fetching Sample open orders', params)

  const { credentials } = exchange

  const { http = new SampleHttp() } = params

  const rawOrders = await http.authedRequest<ISampleOrderSchema[]>({
    verb: AlunaHttpVerbEnum.GET,
    url: `${SAMPLE_PRODUCTION_URL}/orders/open`,
    credentials,
  })

  const { requestCount } = http

  return {
    rawOrders,
    requestCount,
  }

}
